hdf5 1.8.5-r19586 (reduced)
This tree was extracted from upstream hdf5 by the following shell code.
This revision coinsides with the 1_8_5-patch1 svn tag.

    url=https://svn.hdfgroup.uiuc.edu/hdf5/branches/hdf5_1_8_cmake
    v=1.8.5
    r=19586
    paths="
      ACKNOWLEDGMENTS
      CTestConfig.cmake
      README.txt
      config/cmake
      CMakeLists.txt
      COPYING
      src
      c++/CMakeLists.txt
      c++/COPYING
      c++/src
      hl/CMakeLists.txt
      hl/COPYING
      hl/src
      hl/c++/CMakeLists.txt
      hl/c++/COPYING
      hl/c++/src
    "
    date=$(svn log -q -c$r $url |
           sed -n "/^r/ {s/[^|]*|[^|]*|//;p;}")
    svn export -r$r $url hdf5-$v-r$r &&
    mkdir hdf5-$v-r$r-reduced &&
    (cd hdf5-$v-r$r && tar c $paths) |
    (cd hdf5-$v-r$r-reduced && tar x)
    echo "r$r date: $date"

+++ src/H5SMmessage.c
@@ -309,7 +309,7 @@ H5SM_message_encode(uint8_t *raw, const void *_nrecord, void *_ctx)
         *raw++ = 0;     /* reserved (possible flags byte) */
         *raw++ = (uint8_t)message->msg_type_id;
         UINT16ENCODE(raw, message->u.mesg_loc.index);
-        H5F_addr_encode_len(ctx->sizeof_addr, &raw, message->u.mesg_loc.oh_addr);
+        H5F_addr_encode_len((size_t)ctx->sizeof_addr, &raw, message->u.mesg_loc.oh_addr);
     } /* end else */
 
     FUNC_LEAVE_NOAPI(SUCCEED)
@@ -350,7 +350,7 @@ H5SM_message_decode(const uint8_t *raw, void *_nrecord, void *_ctx)
         raw++;          /* reserved */
         message->msg_type_id = *raw++;
         UINT16DECODE(raw, message->u.mesg_loc.index);
-        H5F_addr_decode_len(ctx->sizeof_addr, &raw, &message->u.mesg_loc.oh_addr);
+        H5F_addr_decode_len((size_t)ctx->sizeof_addr, &raw, &message->u.mesg_loc.oh_addr);
     } /* end else */
 
     FUNC_LEAVE_NOAPI(SUCCEED)
